deploy: move to own package
